python/requirements.txt
numpy
pandas
matplotlib
opencv-contrib-python
fastapi
uvicorn
sqlalchemy
pillow
psycopg2
pytz
dlib
cmake
face-recognition
python-multipart